TODO:
HGMDPModel::next produces a new state from an older state and an action applied to it.
The new state could be synonim of a pre-existing one. It is detected after
creating the new MDPState and checking the == operator another similar one
exists. Once this occurs the new MDPState must be destroyed and the found MDPState
should be referenced from the older MDPState.

Right now that feature is not coded because wanted to see if UCT already manages it.