Reducing Stochasticity :
added:
MoveHomeAction:equal compares LR locations instead of HR locations
ForageAction::equal compares categorized reduced biomass features
(maybe must not be included in the comparison, only compare if
go to the same sector to forage, any reward is reflexed in the MDPState).

doing: The half-forage for the move home happens
in the same sector where the agent moves to.