New Parameter for agent : initialPosition. If any component of the Point2D is <0 then GujaratWorld will call the method to assign a random sound coordinate for the agent to settle; otherwise "initialPosition" will be assigned.
MoveHomeAction generates actions that aim to a LRCell with a maximal amount of
resource M. All candidate LRCells involved in the generation of each candidate
action to explore by UCT achieves this bound. A Sector can only yield 0 or 1
candidate for the candidates list.

This way an MDPState will store for exploring 8 ForageActions and at most 8 MoveHomeActions (8 comes
from number of sectors, 8 are defined for the agent)